Gui: Change SelectionFilterPy to new PyCXX extension type
